+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 * fix: static analysis
 * refactor: refactor proxy, introduce proxy metadata
 * feat: PHP 8.4 lazy object
+* refactor(proxy): remove deprecated classes, move dynamic properties checking to
+  `DynamicPropertiesProxyFactory`
 
 ## 1.14.1